Add an iframe title, reusing the linktitle message.
When evaluating wiki pages using EmbedVideo for accessibility using ANDI
(https://www.ssa.gov/accessibility/andi/help/install.html), it noted
that the iframes being created didn't have title elements and that was
an accessibility concern.  I left a place for the iframetitle to be
made an configurable option at a further date, and reused the already available
'ev_default_play_desc' in the case it wasn't set.