Remove `getLocalRefPath()` from VideoTransformOutput and AudioTransformOutput in class initialization
If `S3` or other external storage provider is used to store files in mediawiki, this calls forces download of this files to local storage even if they will be not used to make transformations. This can be very costly operation if used to list multiple files on eg. Category page.